backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn==0.34.2
python-multipart==0.0.21

# Database
firebase-admin==7.1.0

# Image Processing & OCR
pillow==12.0.0
pytesseract==0.3.13
pdf2image==1.17.0

# HTTP Client (for Gemini API)
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6

# Testing
requests==2.32.5
